+++ packages/data-adapters/src/index.ts
@@ -13,6 +13,7 @@ export * from './Provider';
 export * from './MockProvider';
 export * from './BinanceProvider';
 export * from './FrankfurterProvider';
+export * from './DexscreenerProvider';
 export * from './providerRegistry';
 export * from './conformance';
 export { YahooProvider } from './stubs/YahooProvider';